feat(main): add dynamic UI elements
This commit adds dynamic UI elements to the app, allowing for more information to be displayed to the user. Specifically, a new stack of widgets has been added that displays the current connection state of the app, as well as a new container that appears when the app is not connected. These changes were made to improve the user experience and provide more information about the app's current state.