ci(codemagic): Add codemagic CI/CD configuration file
The codemagic.yaml file contains the configuration for two workflows: googleplay-internal and testFlight. These workflows define the build and publishing steps for the internal and beta versions of the app. This commit adds the codemagic configuration file to the repository to automate the build and publishing process for the app.